refactor(web): repository.ts を lib/api 配下に移す
src 直下から lib/api/client.ts を import する配置が API 層の依存方向と逆転していたため lib/api/repository.ts に移動する。client と同階層になり import も "./client" で済むようになる。consumer 側は "~/repository" → "~/lib/api/repository" に追従。
